requirements.txt
garminconnect==0.2.26
logger_setup==0.17
mysql_connector_repackaged==0.3.1
python-dotenv==1.1.0
pytz==2025.1
